requirements.txt
# Cascade Linter GUI - Enhanced Edition v2.0
# Core GUI Framework
PySide6>=6.5.0

# Terminal UI and Logging
rich>=13.0.0
structlog>=23.0.0

# Linting Tools
ruff>=0.1.0
flake8>=6.0.0
mypy>=1.0.0
bandit>=1.7.0

# Auto-fixing Tools
autopep8>=2.0.0
autoflake>=2.0.0

# Optional Enhancements
qtawesome>=1.2.0  # Icons
qt-material>=2.14  # Material Design themes
QT-PyQt-PySide-Custom-Widgets>=1.0.0  # Advanced custom widgets
pytz>=2023.3      # Timezone support

# Tier 2 CLI Features
watchfiles>=0.20.0  # For watch mode

# Development Dependencies (optional)
pytest>=7.0.0
black>=23.0.0
isort>=5.0.0
